+++ src/constants.ts
@@ -230,6 +230,9 @@ export const TEST_IDS = {
     fieldTime: createSelector('data-testid nested-object-cards-editor field-time'),
     fieldAuthor: createSelector('data-testid nested-object-cards-editor field-author'),
     fieldBody: createSelector('data-testid nested-object-cards-editor field-body'),
+    fieldDisplay: createSelector('data-testid nested-object-cards-editor field-display'),
+    fieldDisplayCount: createSelector('data-testid nested-object-cards-editor field-display-count'),
+    option: createSelector((name: unknown) => `nested-object-cards-editor options-${name}`),
   },
   nestedObjectCardContent: {
     root: createSelector('data-testid nested-object-card-content'),
@@ -253,6 +256,7 @@ export const TEST_IDS = {
     buttonShowItems: createSelector('data-testid nested-object-cards-control button-show-items'),
     noItemsMessage: createSelector('data-testid nested-object-cards-control no-items-message'),
     buttonCloseDrawer: createSelector(selectors.components.Drawer.General.close),
+    list: createSelector('data-testid nested-object-cards-control list'),
   },
   nestedObjectsEditor: {
     buttonAddNew: createSelector('data-testid nested-objects-editor button-add-new'),